fix: fetch routes data logic
