+++ src/components/page-slice/routes/list.tsx
@@ -1,0 +1,129 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import type { ProColumns } from '@ant-design/pro-components';
+import { ProTable } from '@ant-design/pro-components';
+import { useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
+
+import { type UseListReturn } from '@/apis/hooks';
+import { DeleteResourceBtn } from '@/components/page/DeleteResourceBtn';
+import {
+  ToAddPageBtn,
+  type ToAddPageBtnProps,
+  ToDetailPageBtn,
+  type ToDetailPageBtnProps,
+} from '@/components/page/ToAddPageBtn';
+import { AntdConfigProvider } from '@/config/antdConfigProvider';
+import { API_ROUTES } from '@/config/constant';
+import type { APISIXType } from '@/types/schema/apisix';
+import type { PageSearchType } from '@/types/schema/pageSearch';
+import type { ListPageKeys } from '@/utils/useTablePagination';
+
+type RouteListProps = {
+  detailTo: ToDetailPageBtnProps['to'];
+  addTo: ToAddPageBtnProps['to'];
+  req: UseListReturn<ListPageKeys, PageSearchType, APISIXType['Route']>;
+};
+
+export const RouteList = (props: RouteListProps) => {
+  const { detailTo, addTo, req } = props;
+  const { data, isLoading, refetch, pagination } = req;
+  const { t } = useTranslation();
+
+  const columns = useMemo<ProColumns<APISIXType['RespRouteItem']>[]>(() => {
+    return [
+      {
+        dataIndex: ['value', 'id'],
+        title: 'ID',
+        key: 'id',
+        valueType: 'text',
+      },
+      {
+        dataIndex: ['value', 'name'],
+        title: t('form.basic.name'),
+        key: 'name',
+        valueType: 'text',
+      },
+      {
+        dataIndex: ['value', 'desc'],
+        title: t('form.basic.desc'),
+        key: 'desc',
+        valueType: 'text',
+      },
+      {
+        dataIndex: ['value', 'uri'],
+        title: 'URI',
+        key: 'uri',
+        valueType: 'text',
+      },
+      {
+        title: t('table.actions'),
+        valueType: 'option',
+        key: 'option',
+        width: 120,
+        render: (_, record) => [
+          <ToDetailPageBtn
+            key="detail"
+            to={detailTo}
+            params={{ id: record.value.id }}
+          />,
+          <DeleteResourceBtn
+            key="delete"
+            name={t('routes.singular')}
+            target={record.value.id}
+            api={`${API_ROUTES}/${record.value.id}`}
+            onSuccess={refetch}
+          />,
+        ],
+      },
+    ];
+  }, [t, refetch, detailTo]);
+
+  return (
+    <AntdConfigProvider>
+      <ProTable
+        columns={columns}
+        dataSource={data.list}
+        rowKey="id"
+        loading={isLoading}
+        search={false}
+        options={false}
+        pagination={pagination}
+        cardProps={{ bodyStyle: { padding: 0 } }}
+        toolbar={{
+          menu: {
+            type: 'inline',
+            items: [
+              {
+                key: 'add',
+                label: (
+                  <ToAddPageBtn
+                    key="add"
+                    label={t('info.add.title', {
+                      name: t('routes.singular'),
+                    })}
+                    to={addTo}
+                  />
+                ),
+              },
+            ],
+          },
+        }}
+      />
+    </AntdConfigProvider>
+  );
+};
